1_llm_api/5_gemini/1_basics/pyproject.toml
[project]
name = "1-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.21.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
